Update ms slugs
Note: I think its a good idea to make the slugs a bit longer to make it
easier to make them unique. Now I have composed the slugs of six
characters from the manuscript location (favouring colleges over towns
when relevant) and the ms number.
This approach could be systematized.

Another possibility would be to use the last 8 characters of the SHA-1
hash of the ms. name (town, library, shelfmark, number). But this might
work for now.